using conftest and adding test_enabled_features
